+++ jsonschema/validators.py
@@ -498,7 +498,47 @@ Draft6Validator = create(
     version="draft6",
 )
 
-_LATEST_VERSION = Draft6Validator
+Draft7Validator = create(
+    meta_schema=_utils.load_schema("draft7"),
+    validators={
+        u"$ref": _validators.ref,
+        u"additionalItems": _validators.additionalItems,
+        u"additionalProperties": _validators.additionalProperties,
+        u"allOf": _validators.allOf,
+        u"anyOf": _validators.anyOf,
+        u"const": _validators.const,
+        u"contains": _validators.contains,
+        u"dependencies": _validators.dependencies,
+        u"enum": _validators.enum,
+        u"exclusiveMaximum": _validators.exclusiveMaximum,
+        u"exclusiveMinimum": _validators.exclusiveMinimum,
+        u"format": _validators.format,
+        u"if": _validators.if_,
+        u"items": _validators.items,
+        u"maxItems": _validators.maxItems,
+        u"maxLength": _validators.maxLength,
+        u"maxProperties": _validators.maxProperties,
+        u"maximum": _validators.maximum,
+        u"minItems": _validators.minItems,
+        u"minLength": _validators.minLength,
+        u"minProperties": _validators.minProperties,
+        u"minimum": _validators.minimum,
+        u"multipleOf": _validators.multipleOf,
+        u"oneOf": _validators.oneOf,
+        u"not": _validators.not_,
+        u"pattern": _validators.pattern,
+        u"patternProperties": _validators.patternProperties,
+        u"properties": _validators.properties,
+        u"propertyNames": _validators.propertyNames,
+        u"required": _validators.required,
+        u"type": _validators.type,
+        u"uniqueItems": _validators.uniqueItems,
+    },
+    type_checker=_types.draft7_type_checker,
+    version="draft7",
+)
+
+_LATEST_VERSION = Draft7Validator
 
 
 class RefResolver(object):
@@ -775,7 +815,7 @@ def validate(instance, schema, cls=None, *args, **kwargs):
     in less obvious or consistent ways. If you know you have a valid schema
     already or don't care, you might prefer using the
     `IValidator.validate` method directly on a specific validator
-    (e.g. ``Draft6Validator.validate``).
+    (e.g. ``Draft7Validator.validate``).
 
 
     Arguments:
@@ -798,7 +838,7 @@ def validate(instance, schema, cls=None, *args, **kwargs):
     proper validator will be used.  The specification recommends that all
     schemas contain :validator:`$schema` properties for this reason. If no
     :validator:`$schema` property is found, the default validator class is
-    `Draft6Validator`.
+    the latest released draft.
 
     Any other provided positional and keyword arguments will be passed on when
     instantiating the ``cls``.